Create navbar component & mount in App component
Store in components folder within src folder.

Add Tailwind CSS styling to the navbar.

Create a theme state to toggle between light and dark themes.

Pass theme values returned by useState as props to the navbar component.

Conditionally render the logo image of the navbar based on the current
theme which was passed as a prop into the component.