Add menu items to navbar component & add sidebar
Add menu items to the navbar component and style them using
Tailwind CSS.

These menu items are links to Home, Expertise, Our Approach, Contact Us
as well as the CTA button.

Add sidebar state and display the sidebar. Have button to leave sidebar
and to open it up.

Include onclick attributes to be able to close sidepanel when either
the x is clicked or menu items are clicked.